refactor: delete deprecated files
