--- src/enums.ts
@@ -1,11 +1,0 @@
-export const STATUS_CODES = {
-    CREATED: 201,
-    NO_CONTENT: 204,
-    BAD_REQUEST: 400,
-    NOT_FOUND: 404,
-    INTERNAL_SERVER_ERROR: 500,
-}
-
-export const ORDERED_BY = {
-    APPOINTMENTS_COUNT: 'appointmentsCount',
-}
