--- src/error-messages.ts
@@ -1,10 +1,0 @@
-export const ERRORS = {
-    ENTITY_ALREADY_EXISTS: 'Entity [%s] already exists',
-    INVALID_PHONE_FORMAT: 'Phones [%s] format, does not follow E.164 standart',
-    MISSING_PARAMETER: 'Missing [%s]',
-    INVALID_DATE_FORMAT: 'Date  format does not follow ISO standart',
-    PAST_DATE: 'Date [%s] can not be in the past',
-    DOCTOR_NOT_AVAILABLE: 'Doctor [%s] is not available at [%s]',
-    ENTITY_NOT_EXISTS: 'Entity [%s] not exists',
-    PHONE_IS_TAKEN: 'Phone [%s] is allready taken',
-}
