+++ src/errors/appointment-conflict-error.ts
