--- src/exceptions/missing-parameter-error.ts
@@ -1,6 +1,0 @@
-export class MissingParameterError extends Error {
-    constructor(message) {
-      super(message);
-      Error.captureStackTrace(this, MissingParameterError);
-    }
-}
