--- src/regular-expressions.ts
@@ -1,4 +1,0 @@
-export const REGEXPRESSIONS = {
-    PHONE_NUMBER: /^\+[1-9]\d{1,14}$/,
-    ISO_DATE: /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2})(\.\d{1,3})?([+-]\d{2}:\d{2}|Z)?$/,
-}
