+++ src/services/patients-service.ts
@@ -1,5 +1,5 @@
 import PatientModel from "../models/patient-model";
-import { DuplicateEntityError } from "../exceptions/duplicate-entity-error";
+import { DuplicateEntityError } from "../errors/duplicate-entity-error";
 import PatientsRepository from "../repositories/patients-repository";
 import CreatePatientDto from "../dto/patients/create-patient-dto";
 import { v4 } from "uuid";
